build: cleans up pending transaction
