+++ packages/common/mui/src/index.ts
@@ -12,6 +12,7 @@ export * from './lib/card-variants';
 export * from './lib/chart-tool-tip/chart-tool-tip';
 export * from './lib/chart-header/chart-header';
 export * from './lib/chart-info-box/chart-info-box';
+export * from './lib/copy-caption/copy-caption';
 export * from './lib/count-up/count-up';
 export * from './lib/countdown-cards/countdown-cards';
 export * from './lib/currency-img/currency-img';
