+++ packages/common/mui/src/lib/copy-caption/copy-caption.tsx
@@ -1,0 +1,35 @@
+import { SxProps, useTheme } from '@mui/material';
+import { ReactNode } from 'react';
+import { Caption } from '../typography/typography';
+import { FormattedMessage } from 'react-intl';
+
+/* eslint-disable-next-line */
+export interface CopyCaptionProps {
+  showAlert: boolean;
+  title?: ReactNode;
+  sx?: SxProps;
+}
+
+export function CopyCaption({ showAlert, title, sx }: CopyCaptionProps) {
+  const theme = useTheme();
+
+  return (
+    <Caption
+      sx={{
+        background: theme.palette.background.paper,
+        color: theme.palette.typography.main,
+        padding: theme.spacing(1.5),
+        position: 'absolute',
+        borderRadius: theme.shape.borderRadius(),
+        border: `1px solid ${theme.palette.primary.light}`,
+        transition: 'all 0.3s ease-in-out',
+        opacity: showAlert ? 1 : 0,
+        ...sx,
+      }}
+    >
+      {title || <FormattedMessage defaultMessage="Address Copied" />}
+    </Caption>
+  );
+}
+
+export default CopyCaption;
